+++ packages/waggle/src/socket/DataServer.ts
@@ -1,7 +1,9 @@
 import express from 'express'
 import { createServer, Server } from 'http'
+import logger from '../logger'
 // eslint-disable-next-line
 const socketio = require('socket.io')
+const log = logger('socket')
 
 export class DataServer {
   public PORT: number
@@ -22,7 +24,7 @@ export class DataServer {
   public listen = async (): Promise<void> => {
     return await new Promise((resolve) => {
       this.server.listen(this.PORT, () => {
-        console.debug(`Server running on port ${this.PORT}`)
+        log(`Data server running on port ${this.PORT}`)
         resolve()
       })
     })