+++ packages/waggle/src/socket/events/certificates.ts
@@ -1,7 +1,9 @@
 import { EventTypesResponse } from '../constantsReponse'
+import logger from '../../logger'
+const log = logger('socket')
 
 export const loadCertificates = (socket: SocketIO.Server, certificates: string[]) => {
-  // console.log(`Sending back ${certificates.length} certificates`)
+  log(`Sending back ${certificates.length} certificates`)
   socket.emit(EventTypesResponse.RESPONSE_GET_CERTIFICATES, {
     certificates
   })