+++ packages/waggle/src/storage/storage.test.ts
@@ -2,9 +2,9 @@ import fs from 'fs'
 import path from 'path'
 import PeerId from 'peer-id'
 import { Config } from '../constants'
-import { createLibp2p, createTmpDir, TmpDir, tmpZbayDirPath, dataFromRootPems } from '../testUtils'
+import { createLibp2p, createTmpDir, TmpDir, tmpZbayDirPath, dataFromRootPems } from '../common/testUtils'
 import { Storage } from './storage'
-import * as utils from '../utils'
+import * as utils from '../common/utils'
 import { createUserCsr, createUserCert, configCrypto } from '@zbayapp/identity'
 jest.setTimeout(30_000)